% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loops.R
\name{check_cycle_pedigree}
\alias{check_cycle_pedigree}
\title{Wrapper Function To Check Pedigrees For Cycles}
\usage{
check_cycle_pedigree(
  ps_pedig_path = NULL,
  pb_report_cycle = FALSE,
  ps_delim = "|",
  pn_id_col = 1,
  pn_sire_col = 2,
  pn_dam_col = 3,
  pcol_types = NULL,
  ptbl_pedigree = NULL
)
}
\arguments{
\item{ps_pedig_path}{path to the pedigree input file}

\item{pb_report_cycle}{flag, whether animals in cycle are reported}

\item{ps_delim}{delimiting character of columns in input file}

\item{pn_id_col}{column index of animals in input file}

\item{pn_sire_col}{column index of sires in input file}

\item{pn_dam_col}{column indes of dams in input file}

\item{pcol_types}{data types of columns in input file, passed to readr}

\item{ptbl_pedigree}{pedigree given as tbl}
}
\description{
The PedigreeCycleCheck R6 class is used to find cycles
in a pedigree. The pedigree can be specified via input file
or via a tibble. The result consists either of TRUE or
FALSE which tells whether a pedigree contains a loop or not.
In addition the animals that are part of the loop can be
listed.
}
\details{
Cycles are searched using depth-first-traversals of the pedigree.
}
\examples{
\dontrun{
check_cycle_pedigree(ps_pedig_path = system.file('extdata','data_sample2.csv', package = 'qprppedigree'))
}

}
