% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedigree_id.R
\name{check_pedig_id}
\alias{check_pedig_id}
\title{Check Animal IDs in Pedigrees}
\usage{
check_pedig_id(
  ps_pedig_path,
  ps_delim = "|",
  ps_id_col = "#IDTier",
  pcol_types = NULL,
  ptbl_pedigree = NULL
)
}
\arguments{
\item{ps_pedig_path}{path to the pedigree input file}

\item{ps_delim}{column delimiting character}

\item{ps_id_col}{column title where animal ID is contained}

\item{pcol_types}{column types of pedigree in ps_pedig_path}

\item{ptbl_pedigree}{tibble containing the pedigree, obtained by read_prp_pedigree()}
}
\value{
list of pedigree characteristics
}
\description{
The software system PopReport analyses pedigrees. The pedigrees used as input
must satisfy certain criteria. This function checks the required properties
of animal IDs in a pedigree. Animal IDs in a pedigree are the primary keys
that identify the individuals.

The main property of the animal IDs consists of their uniqueness. The output
of this function shows potential duplicate IDs.
}
\examples{
\dontrun{
check_pedig_id(ps_pedig_path = system.file('extdata',
    'PopReport_SN_ohne_20210115.csv_adaptfin2.csv',
  package = 'qprppedigree'))
}

}
