% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedigree_datatypes.R
\name{check_pedigree_datatypes}
\alias{check_pedigree_datatypes}
\title{Checking Datatypes of Columns of a Pedigree}
\usage{
check_pedigree_datatypes(
  ps_pedig_path,
  ps_delim = "|",
  ps_id_col = "#IDTier",
  pl_dtype = NULL,
  pcol_types = NULL,
  ptbl_pedigree = NULL
)
}
\arguments{
\item{ps_pedig_path}{path to the pedigree input file}

\item{ps_delim}{column delimiting character}

\item{ps_id_col}{column title for animal IDs}

\item{pl_dtype}{list of column names with required data-types}

\item{pcol_types}{column types of pedigree in ps_pedig_path used by read_prp_pedigree}

\item{ptbl_pedigree}{tibble containing pedigree information}
}
\value{
list with datatype problems
}
\description{
The datatypes given in pl_dtype$dtp are compared to the actual datatypes
in the pedigree for the columns given in pl_dtype$col. If the datatypes
do not match, then the records causing the problems are returned in the
result list as a tibble.
}
\details{
The format of the argument pl_dtype can be determined from the result of
the function get_pedigree_datatypes().
}
\examples{
\dontrun{
check_pedigree_datatypes(ps_pedig_path = system.file('extdata',
    'PopReport_SN_ohne_20210115.csv_adaptfin2.csv',
  package = 'qprppedigree'))
}

}
