% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pedigree_report.R
\name{create_pedigree_report}
\alias{create_pedigree_report}
\title{Create Pedigree Report}
\usage{
create_pedigree_report(
  ps_pedigree_path,
  ps_report_outdir = ".",
  ps_report_rmd = paste0(format(Sys.time(), "\%Y\%m\%d\%H\%M\%S"),
    "_qpdt_pedigree_report.Rmd", collapse = ""),
  pl_params = NULL
)
}
\arguments{
\item{ps_pedigree_path}{path to the pedigree file}

\item{ps_report_outdir}{report output directory}

\item{ps_report_rmd}{name of the pedigree report file}

\item{pl_params}{list of parameters to be included in the report}
}
\description{
The checks in this package are done for a given pedigree given by ps_pedigree_path.
The results of the checks are summarised in a report which is generated by this
function.
}
\examples{
\dontrun{
s_pedigree_path <- file.path(here::here(), 'inst', 'extdata', 'PopReport_OST_mit_20210318.csv_adaptfin5.csv')
create_pedigree_report(ps_pedigree_path = system.file('extdata',
    'PopReport_SN_ohne_20210115.csv_adaptfin2.csv',package = 'qpdt'),
  ps_report_outdir = file.path(here::here(), 'scratch', 'report_out'))
}

}
