% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_settings.R
\name{get_argus_prp_settings}
\alias{get_argus_prp_settings}
\title{Settings for PopReport Pedigrees Exported From ARGUS}
\usage{
get_argus_prp_settings(
  ps_id_col = "#IDTier",
  ps_sire_col = "IDVater",
  ps_dam_col = "IDMutter",
  ps_bd_col = "Birthdate",
  ps_sex_col = "Geschlecht",
  ps_plz_col = "PLZ",
  ps_introg_col = "introg",
  ps_inb_col = "inb_gen",
  ps_cryo_col = "cryo",
  ps_col_delim = "|"
)
}
\arguments{
\item{ps_id_col}{column header for identifiers}

\item{ps_sire_col}{column header for sire}

\item{ps_dam_col}{column header for dam}

\item{ps_bd_col}{column header for birthday}

\item{ps_sex_col}{column header for sex}

\item{ps_plz_col}{column header for postal code}

\item{ps_introg_col}{column header for introgression}

\item{ps_inb_col}{column header for inbreeding}

\item{ps_cryo_col}{column header for cryo conservation}

\item{ps_col_delim}{column delimiter}
}
\description{
All input parameters that change with the type of pedigree are
returned in a list
}
