% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_settings.R
\name{get_generic_settings}
\alias{get_generic_settings}
\title{Settings for Generic Pedigree}
\usage{
get_generic_settings(
  ps_id_col = "id",
  ps_sire_col = "sire",
  ps_dam_col = "dam",
  ps_bd_col = "birthdate",
  ps_sex_col = "sex",
  ps_plz_col = "plz",
  ps_introg_col = "introg",
  ps_inb_col = "inb_gen",
  ps_cryo_col = "cryo",
  ps_col_delim = "|",
  pl_dtype = list(col = c("idanimal", "idsire", "iddam", "birthdate", "sex", "plz",
    "introg", "inb_gen", "cryo"), dtp = c("integer", "character", "integer", "date",
    "character", "integer", "double", "double", "logical"))
)
}
\arguments{
\item{ps_id_col}{column header for identifiers}

\item{ps_sire_col}{column header for sire}

\item{ps_dam_col}{column header for dam}

\item{ps_bd_col}{column header for birthday}

\item{ps_sex_col}{column header for sex}

\item{ps_plz_col}{column header for postal code}

\item{ps_introg_col}{column header for introgression}

\item{ps_inb_col}{column header for inbreeding}

\item{ps_cryo_col}{column header for cryo conservation}

\item{ps_col_delim}{column delimiter}

\item{pl_dtype}{list with datatypes}
}
\description{
All input parameters that change with the type of pedigree are
returned in a list
}
