% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_settings.R
\name{get_gnm_prp_settings}
\alias{get_gnm_prp_settings}
\title{Settings for PopReport Pedigrees From Genmon}
\usage{
get_gnm_prp_settings(
  ps_id_col = "#animal",
  ps_sire_col = "sire",
  ps_dam_col = "dam",
  ps_bd_col = "birth_date",
  ps_sex_col = "sex",
  ps_plz_col = "plz",
  ps_introg_col = "introg",
  ps_inb_col = "inb_gen",
  ps_cryo_col = "cryo",
  ps_col_delim = "|"
)
}
\arguments{
\item{ps_id_col}{column header for identifiers}

\item{ps_sire_col}{column header for sire}

\item{ps_dam_col}{column header for dam}

\item{ps_bd_col}{column header for birthday}

\item{ps_sex_col}{column header for sex}

\item{ps_plz_col}{column header for postal code}

\item{ps_introg_col}{column header for introgression}

\item{ps_inb_col}{column header for inbreeding}

\item{ps_cryo_col}{column header for cryo conservation}

\item{ps_col_delim}{column delimiter}
}
\description{
All input parameters that change with the type of pedigree are
returned in a list
}
