% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedigree_datatypes.R
\name{get_pedigree_datatypes}
\alias{get_pedigree_datatypes}
\title{Determine Datatypes of Columns in a Pedigree}
\usage{
get_pedigree_datatypes(
  ps_pedig_path,
  ps_delim = "|",
  pcol_types = NULL,
  ptbl_pedigree = NULL
)
}
\arguments{
\item{ps_pedig_path}{path to pedigree input file}

\item{ps_delim}{column delimiter in pedigree input file}

\item{pcol_types}{column types in the format of cols() used by read_prp_pedigree}

\item{ptbl_pedigree}{a tibble containing a pedigree}
}
\value{
list with datatypes
}
\description{
The datatypes of the columns are determined using the function readr::guess_parser().
}
\details{
The result is returned as a list that can be used as input for the function
check_pedigree_datatypes
}
\examples{
\dontrun{
get_pedigree_datatypes(ps_pedig_path = system.file('extdata',
    'PopReport_SN_ohne_20210115.csv_adaptfin2.csv',
  package = 'qprppedigree'))
}

}
