% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_qpdt_logger}
\alias{get_qpdt_logger}
\title{Create log4r Logger for package}
\usage{
get_qpdt_logger(ps_logfile = "qpdt.log", ps_level = "FATAL")
}
\arguments{
\item{ps_logfile}{name of the logfile}

\item{ps_level}{logger level}
}
\value{
qpdt_logger
}
\description{
Create log4r Logger for package
}
\examples{
\dontrun{
qpdt_logger <- get_qpdt_logger()
}
}
