% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedigree_parent.R
\name{get_tbl_parent_sex_err}
\alias{get_tbl_parent_sex_err}
\title{Check Sex of a Parent}
\usage{
get_tbl_parent_sex_err(
  ptbl_pedigree,
  psym_animal_id,
  psym_parent_id,
  psym_sex_col,
  pvec_parent_by,
  ps_wrong_sex
)
}
\arguments{
\item{ptbl_pedigree}{tibble containing pedigree}

\item{psym_animal_id}{symbol for column header of animal ID}

\item{psym_parent_id}{symbol for column header of parent ID}

\item{psym_sex_col}{symbol for column header of sex}

\item{pvec_parent_by}{vector used as by argument in join}

\item{ps_wrong_sex}{value for the wrong sex}
}
\description{
Using inner_joins to determine sex of parents. Filter out the parents with
the wrong sex. The result contains the ID of the animal, the parent and the
wrong sex.
}
