% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qpdt_log_debug}
\alias{qpdt_log_debug}
\title{Wrapper for log4r debug}
\usage{
qpdt_log_debug(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_caller}{function from which we are called}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r debug
}
\examples{
\dontrun{
qpdt_logger <- get_qpdt_logger()
qpdt_log_level(qpdt_logger, 'DEBUG')
qpdt_log_debug(qpdt_logger, 'Examples', 'test message')
}
}
