% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qpdt_log_level}
\alias{qpdt_log_level}
\title{Wrapper to set the level of a logger}
\usage{
qpdt_log_level(
  plogger,
  ps_level = c("DEBUG", "INFO", "WARN", "ERROR", "FATAL")
)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_level}{new level of plogger}
}
\description{
Wrapper to set the level of a logger
}
\examples{
\dontrun{
qpdt_logger <- get_qpdt_logger()
qpdt_log_level(qpdt_logger, 'INFO')
}
}
