% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedigree_helper.R
\name{read_prp_pedigree}
\alias{read_prp_pedigree}
\title{Read Pedigree from csv-file}
\usage{
read_prp_pedigree(ps_pedig_path, ps_delim = "|", pcol_types = NULL)
}
\arguments{
\item{ps_pedig_path}{path to the pedigree input file}

\item{ps_delim}{column delimiting character}
}
\value{
tbl_pedigree tibble containing pedigree
}
\description{
The pedigree is read from a file with a tabular structure. The columns are
separated by a character given by ps_delim.
}
\examples{
\dontrun{
p <- read_prp_pedigree(ps_pedig_path = system.file('extdata',
'PopReport_SN_ohne_20210115.csv_adaptfin2.csv', package = 'qprppedigree'))
}
}
